# openapi/sovereign.v1.yaml
openapi: 3.1.0
info:
  title: Sovereign REST API
  version: "1.0.0"
  description: |
    Control-plane API for Sovereign tenants. Hand-written source of truth
    (D-5): the route set in internal/api/router.go must stay in parity with
    this spec in both directions, enforced by the OpenAPI drift test.
servers:
  - url: https://id.<domain>/api/v1
    description: Identity host, versioned API base.
# Default security for the API: bearer (programmatic) or session cookie
# (browser). Operations that are anonymous override this with `security: []`.
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /me/credentials:
    get:
      operationId: listSelfCredentials
      summary: List the current principal's WebAuthn credentials
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        "200":
          description: WebAuthn credentials
          content:
            application/json:
              schema:
                type: array
                maxItems: 100
                items:
                  $ref: "#/components/schemas/WebAuthnCredential"
        "401":
          $ref: "#/components/responses/Problem"
  /me/credentials/{id}:
    delete:
      operationId: deleteSelfCredential
      summary: Delete one WebAuthn credential (409 when it is the last passkey)
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Credential deleted
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
...
  /admin/users/{id}/credentials:
    get:
      operationId: adminListUserCredentials
      summary: List a user's WebAuthn credentials (scope admin:users:read)
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of passkeys
          content:
            application/json:
              schema:
                type: array
                maxItems: 100
                items:
                  $ref: "#/components/schemas/WebAuthnCredential"
        "404":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  responses:
    Problem:
      description: RFC 9457 problem details
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    Problem:
      type: object
      description: RFC 9457 problem details
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri
        errors:
          type: array
          maxItems: 10
          items:
            $ref: "#/components/schemas/FieldError"
    FieldError:
      type: object
      required: [field, code]
      properties:
        field:
          type: string
        code:
          type: string
        detail:
          type: string
...
    WebAuthnCredential:
      type: object
      description: A registered passkey. Public key material and attestation data are never served.
      properties:
        id:
          type: string
        credential_id:
          type: string
        created_at:
          type: string
          format: date-time
        last_used_at:
          type: string
          format: date-time
...